magit-ediff-compare--read-files: allow RANGE and REVA to be nil
That was the intention in the first place.  When they (we only check
RANGE though) are nil, then that signals that the user did something
that should result in the changes in a certain commit to be displayed.
That is done by comparing <commit> with <commit>^, or
<commit>^..commit>, but when we then go on asking for a file to compare,
then we shouldn't be talking about a range, but a single commit.

RANGE being nil is used to signal that.  But there was a bug, the RANGE
is the only thing being passed on to `magit-read-changed-file', so now
we have to construct it from REVB only, or if REVA is non-nil, from that
too.