magit-ediff-compare: Change single rev handling
When given a nil revision, magit-ediff-compare diffs with the working
tree (like "git diff" does).  However, when
magit-ediff-compare--read-revisions reads in revisions, a revision for
is converts to "revB^..revB", so there is no way to interactively
compare with the working tree.

Change magit-ediff-compare--read-revisions to allow a nil revision.  No
longer return a range with magit-ediff-compare--read-revisions because
this information is not needed once a "..." range is detected and revA
is adjusted to the merge-base.

Update magit-ediff-compare--read-files to handle the working tree
comparison (i.e., a nil revision).